requirements.txt
redis
pyyaml
flask
uvicorn[standard]
sqlalchemy
requests
beautifulsoup4
html2text
python-magic
firecrawl-py